# api/openapi.yaml
openapi: 3.1.0
info:
  title: Orchestrator API (единый API всех каналов)
  version: 0.1.0-mvp
  description: |
    Единый API оркестратора. Через него работают все каналы (web PWA, Telegram-бот).
    Машинный канал (агент ↔ оркестратор) идёт по WebSocket с аутентификацией по UUID
    и здесь описан только как точка подключения — формат сообщений см. docs/protocol.md.
    Трассировка требований: коды FR (A1, B2, E2…) и Gherkin-функции даны в описаниях.
servers:
  - url: https://api.example.com
    description: prod (заменить на свой домен)
security:
  - bearerAuth: []
...
paths:
...
  /integrations/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [integrations]
      summary: Получить интеграцию
      responses:
        "200":
          description: Интеграция
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrationWithSecret" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [integrations]
      summary: Редактировать имя/IP
      description: Не прерывает активные задачи (FR B5, Gherkin §2).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntegrationUpdate" }
      responses:
        "200": { description: Обновлено }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [integrations]
      summary: Удалить интеграцию
      description: |
        При наличии активных задач требует подтверждения (?confirm=true) и корректно
        завершает/отменяет их (FR B5, Gherkin §2 «Удаление требует подтверждения»).
      parameters:
        - name: confirm
          in: query
          schema: { type: boolean, default: false }
      responses:
        "204": { description: Удалено }
        "409":
          description: Есть активные задачи, нужно подтверждение (confirm=true)

  # ---------- TASKS (FR E,F,H; Gherkin §4–§10) ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    IntegrationUpdate:
      type: object
      properties:
        name: { type: string }
        ip_hint: { type: string, nullable: true }
    Integration:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        ip_hint: { type: string, nullable: true }
        status: { type: string, enum: [online, offline], description: FR B4 }
        last_seen_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    IntegrationWithSecret:
      allOf:
        - $ref: "#/components/schemas/Integration"
        - type: object
          properties:
            uuid:
              type: string
              format: uuid
              description: Секрет для аутентификации машины (FR B2). Показывается владельцу.
...