# orchestrator/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

// Привязка внешнего канала (Telegram) к аккаунту (FR D3).
type ChannelLink struct {
	ID         pgtype.UUID        `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	Channel    string             `json:"channel"`
	ExternalID string             `json:"external_id"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

// Одноразовые коды привязки канала (FR D3).
type ChannelLinkCode struct {
	Code      string             `json:"code"`
	UserID    pgtype.UUID        `json:"user_id"`
	Channel   string             `json:"channel"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

// Машины пользователя. Аутентификация по UUID-секрету (FR B3,B6).
type Integration struct {
	ID     pgtype.UUID `json:"id"`
	UserID pgtype.UUID `json:"user_id"`
	Name   string      `json:"name"`
	IpHint *string     `json:"ip_hint"`
	// HMAC от UUID — для аутентификации машины без хранения секрета в открытом виде.
	UuidHmac string `json:"uuid_hmac"`
	// Зашифрованный UUID — чтобы показать владельцу при настройке (FR B2).
	UuidEnc []byte `json:"uuid_enc"`
	// online/offline по heartbeat через шину, без живого соединения (FR B4).
	Status     string             `json:"status"`
	LastSeenAt pgtype.Timestamptz `json:"last_seen_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
	// NULL = активна; иначе — момент мягкого удаления (ADR 0004). Owner-scoped запросы и аутентификация машины (GetIntegrationByUUIDHMAC) фильтруют deleted_at IS NULL — удалённая интеграция считается несуществующей.
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

// Непрозрачные refresh-токены (хэш). Поддержка logout/refresh (FR A3).
type RefreshToken struct {
	ID        pgtype.UUID        `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	Revoked   bool               `json:"revoked"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

// Секретный токен регистрации. Утечка не критична (FR A2).
type RegistrationToken struct {
	ID        pgtype.UUID        `json:"id"`
	Token     string             `json:"token"`
	IsActive  bool               `json:"is_active"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

// Задачи. Статус — FSM (docs/Жизненный цикл задачи.md). История бессрочна (FR I2).
type Task struct {
	ID            pgtype.UUID `json:"id"`
	UserID        pgtype.UUID `json:"user_id"`
	IntegrationID pgtype.UUID `json:"integration_id"`
	TextEnc       []byte      `json:"text_enc"`
	// completed выставляется ТОЛЬКО явным подтверждением пользователя (FR E2).
	Status         string             `json:"status"`
	IdempotencyKey *string            `json:"idempotency_key"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}

// Журнал событий = история (FR H1) + аудит (FR F4). seq хранит порядок (FR F2).
type TaskEvent struct {
	ID     pgtype.UUID `json:"id"`
	TaskID pgtype.UUID `json:"task_id"`
	Seq    int64       `json:"seq"`
	Type   string      `json:"type"`
	// Ответ/решение ссылается на свой вопрос/запрос (FR F2).
	RefEventID pgtype.UUID        `json:"ref_event_id"`
	PayloadEnc []byte             `json:"payload_enc"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

// Аккаунты. Доступ закрытый, регистрация только по токену (FR A1).
type User struct {
	ID           pgtype.UUID        `json:"id"`
	Username     string             `json:"username"`
	PasswordHash string             `json:"password_hash"`
	IsAdmin      bool               `json:"is_admin"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}